requirements.txt
pelican
fabric
Markdown
IPython
jupyter
beautifulsoup4
ghp-import
py_w3c
embedly-cards
pyembed-markdown
